Add check_idxqt_speed

+++ test_scripts/_run_tests.bash
@@ -9,6 +9,16 @@ if [ x$reseek == x ] ; then
 	exit 1
 fi
 
+if [ ! -d ../big_data ] ; then
+	echo "../big_data not found"
+	exit 1
+fi
+
+if [ `uname -n` != "rip" ] ; then
+	echo Must run on rip
+	exit 1
+fi
+
 rm -rf ../test_output
 mkdir ../test_output
 mkdir -p ../test_results
@@ -22,6 +32,7 @@ git status >> $log
 
 echo STARTED `date` >> $log
 
+./idxqt_speed.bash
 ./convert.bash
 ./align.bash
 ./alignpair.bash
@@ -30,6 +41,7 @@ echo STARTED `date` >> $log
 ./scop40.bash
 ./pdb2mega.bash
 
+python3 ./check_idxqt_speed.py >> $log
 python3 ./check_logs.py >> $log
 python3 ./check_convert.py >> $log
 python3 ./check_columns.py >> $log
